Code refactor of AdobeSignService workflow
