+++ src/main/java/com/imagination/cbs/service/impl/AdobeSignServiceImpl.java
@@ -1,42 +1,24 @@
 package com.imagination.cbs.service.impl;
 
-import static com.imagination.cbs.util.AdobeConstant.ACCESS_TOKEN;
-import static com.imagination.cbs.util.AdobeConstant.ADOBE;
 import static com.imagination.cbs.util.AdobeConstant.ADOBE_ACCESS_TOKEN;
 import static com.imagination.cbs.util.AdobeConstant.ADOBE_ACCESS_TOKEN_EXP_TIME;
 import static com.imagination.cbs.util.AdobeConstant.ADOBE_AUTH_CODE;
-import static com.imagination.cbs.util.AdobeConstant.ADOBE_CLIENT_ID;
-import static com.imagination.cbs.util.AdobeConstant.ADOBE_CLIENT_SECRET;
-import static com.imagination.cbs.util.AdobeConstant.ADOBE_GRANT_TYPE;
-import static com.imagination.cbs.util.AdobeConstant.ADOBE_REDIRECT_URL;
 import static com.imagination.cbs.util.AdobeConstant.ADOBE_REFRESH_TOKEN;
 import static com.imagination.cbs.util.AdobeConstant.ADOBE_TOKEN_TYPE;
 import static com.imagination.cbs.util.AdobeConstant.AGREEMENTS_COMBINEDDOCUMENT;
 import static com.imagination.cbs.util.AdobeConstant.AGREEMENTS_ENDPOINT;
-import static com.imagination.cbs.util.AdobeConstant.BEARER;
-import static com.imagination.cbs.util.AdobeConstant.CLIENT_ID;
-import static com.imagination.cbs.util.AdobeConstant.CLIENT_SECRET;
-import static com.imagination.cbs.util.AdobeConstant.CODE;
 import static com.imagination.cbs.util.AdobeConstant.EMAIL;
-import static com.imagination.cbs.util.AdobeConstant.EXPIRES_IN;
 import static com.imagination.cbs.util.AdobeConstant.FILEINFOS;
-import static com.imagination.cbs.util.AdobeConstant.GRANT_TYPE;
 import static com.imagination.cbs.util.AdobeConstant.ID;
 import static com.imagination.cbs.util.AdobeConstant.MEMBERINFOS;
 import static com.imagination.cbs.util.AdobeConstant.NAME;
-import static com.imagination.cbs.util.AdobeConstant.OAUTH_ACCESS_TOKEN_ENDPOINT;
-import static com.imagination.cbs.util.AdobeConstant.OAUTH_BASE_URL;
-import static com.imagination.cbs.util.AdobeConstant.OAUTH_REFRESH_TOKEN_ENDPOINT;
 import static com.imagination.cbs.util.AdobeConstant.ORDER;
 import static com.imagination.cbs.util.AdobeConstant.PARTICIPANTSETSINFO;
-import static com.imagination.cbs.util.AdobeConstant.REDIRECT_URI;
-import static com.imagination.cbs.util.AdobeConstant.REFRESH_TOKEN;
 import static com.imagination.cbs.util.AdobeConstant.ROLE;
 import static com.imagination.cbs.util.AdobeConstant.SIGNATURETYPE;
 import static com.imagination.cbs.util.AdobeConstant.SIGNATURE_ESIGN;
 import static com.imagination.cbs.util.AdobeConstant.STATE;
 import static com.imagination.cbs.util.AdobeConstant.STATE_IN_PROCESS;
-import static com.imagination.cbs.util.AdobeConstant.TOKEN_TYPE;
 import static com.imagination.cbs.util.AdobeConstant.TRANSIENT_DOCUMENTS_ENDPOINT;
 import static com.imagination.cbs.util.AdobeConstant.TRANSIENT_DOCUMENT_ID;
 
@@ -46,10 +28,9 @@ import java.io.InputStream;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.StringJoiner;
-import java.util.stream.Collectors;
 
 import javax.transaction.Transactional;
 
@@ -64,9 +45,7 @@ import org.springframework.http.HttpMethod;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.util.CollectionUtils;
 import org.springframework.util.LinkedMultiValueMap;
-import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponents;
 import org.springframework.web.util.UriComponentsBuilder;
@@ -81,6 +60,7 @@ import com.imagination.cbs.exception.CBSApplicationException;
 import com.imagination.cbs.exception.ResourceNotFoundException;
 import com.imagination.cbs.repository.ConfigRepository;
 import com.imagination.cbs.service.AdobeSignService;
+import com.imagination.cbs.util.AdobeUtility;
 import com.imagination.cbs.util.AdobeUtils;
 
 import lombok.extern.slf4j.Slf4j;
@@ -98,228 +78,8 @@ public class AdobeSignServiceImpl implements AdobeSignService {
 	@Autowired
 	private Environment env;
 
-	private String getOauthAccessToken() {
-
-		String oauthAccessToken = "";
-		String oauthRefreshToken = "";
-
-		try {
-
-			Map<String, Config> keys = getAdobeKeyDetails(ADOBE);
-
-			if (!isMapKeyValueEmptyOrNull(keys, ADOBE_ACCESS_TOKEN)) {
-
-				if (AdobeUtils.isExpired(keys.get(ADOBE_ACCESS_TOKEN_EXP_TIME).getKeyValue())) {
-
-					oauthAccessToken = keys.get(ADOBE_ACCESS_TOKEN).getKeyValue();
-					return BEARER + oauthAccessToken;
-
-				} else {
-					oauthRefreshToken = keys.get(ADOBE_REFRESH_TOKEN).getKeyValue();
-					oauthAccessToken = getNewAccessToken(oauthRefreshToken).getAccessToken();
-					return BEARER + oauthAccessToken;
-				}
-
-			} else {
-
-				if (!isMapKeyValueEmptyOrNull(keys, ADOBE_REFRESH_TOKEN))
-					oauthRefreshToken = keys.get(ADOBE_REFRESH_TOKEN).getKeyValue();
-
-				oauthAccessToken = getNewAccessToken(oauthRefreshToken).getAccessToken();
-			}
-
-			log.debug("BEARER TOKEN:::: {}", BEARER + oauthAccessToken);
-
-			return BEARER + oauthAccessToken;
-
-		} catch (RuntimeException runtimeException) {
-
-			throw new CBSApplicationException("Adobe Keys not found inside Config table");
-
-		}
-	}
-
-	private AdobeOAuthDto getOauthAccessTokenFromRefreshToken(String oAuthRefreshToken) {
-
-		log.debug("oAuthToken::Refresh:: {}", oAuthRefreshToken);
-		JsonNode response = null;
-		AdobeOAuthDto adobeOAuthDto = new AdobeOAuthDto();
-
-		String uri = OAUTH_BASE_URL + OAUTH_REFRESH_TOKEN_ENDPOINT;
-		HttpHeaders headers = new HttpHeaders();
-
-		try {
-
-			headers.add(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_FORM_URLENCODED_VALUE);
-
-			MultiValueMap<String, String> requestBody = getRequestBody(oAuthRefreshToken);
-			HttpEntity<?> httpEntity = new HttpEntity<>(requestBody, headers);
-
-			response = restTemplate.exchange(uri, HttpMethod.POST, httpEntity, JsonNode.class).getBody();
-			adobeOAuthDto = convertJsonToObj(response);
-			log.info("Get access token using aefresh token::: {}", adobeOAuthDto);
-
-			adobeOAuthDto.setRefreshToken(oAuthRefreshToken);
-
-		} catch (Exception e) {
-			log.info("Exception refresh token:::{}", e);
-		}
-		return adobeOAuthDto;
-	}
-
-	private String getBaseURIForRestAPI(String accessToken) {
-
-		String servicesBaseUrl = "";
-		Map<String, Config> keyValue = getAdobeKeyDetails("ADOBE_SERVICES_BASE_URL");
-
-		log.info("servicesBaseUrl= {}", keyValue);
-
-		if (!isNullOrEmptyMap(keyValue)) {
-
-			servicesBaseUrl = keyValue.get("ADOBE_SERVICES_BASE_URL").getKeyValue();
-			return servicesBaseUrl;
-
-		} else {
-
-			ResponseEntity<JsonNode> res = null;
-
-			HttpHeaders head = new HttpHeaders();
-			head.add(HttpHeaders.AUTHORIZATION, accessToken);
-
-			HttpEntity<String> httpEntity = new HttpEntity<>(head);
-
-			res = restTemplate.exchange(servicesBaseUrl, HttpMethod.GET, httpEntity, JsonNode.class);
-
-			servicesBaseUrl = res.getBody().path("apiAccessPoint").asText() + "api/rest/v6";
-			log.info("ApiAccessPoint BaseUris:::{}", servicesBaseUrl);
-
-			return servicesBaseUrl;
-		}
-	}
-
-	private AdobeOAuthDto convertJsonToObj(JsonNode res) {
-
-		AdobeOAuthDto adobeOAuthDto = new AdobeOAuthDto();
-
-		try {
-			adobeOAuthDto.setAccessToken(res.path(ACCESS_TOKEN).asText());
-			adobeOAuthDto.setRefreshToken(res.path(REFRESH_TOKEN).asText());
-			adobeOAuthDto.setTokenType(res.path(TOKEN_TYPE).asText());
-			adobeOAuthDto.setExpiresIn(res.path(EXPIRES_IN).asInt());
-			log.info("JsonNode to AdobeoAuth:: {}", adobeOAuthDto);
-
-		} catch (Exception e) {
-			log.info("Exception inside convertJsonToObj():: {}" + e);
-		}
-
-		return adobeOAuthDto;
-	}
-
-	private AdobeOAuthDto getNewAccessToken(String oauthRefreshToken) {
-
-		log.info("AdobeOAuth:::OAuthRefreshToken:: {}", oauthRefreshToken);
-
-		ResponseEntity<JsonNode> results = null;
-		AdobeOAuthDto adobeOAuthDto = null;
-
-		String url = OAUTH_BASE_URL + OAUTH_ACCESS_TOKEN_ENDPOINT;
-		MultiValueMap<String, String> headers = new HttpHeaders();
-
-		headers.add(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_FORM_URLENCODED_VALUE);
-		HttpEntity<?> httpEntity = new HttpEntity<>(getRequestBody(), headers);
-
-		try {
-
-			if (oauthRefreshToken.isEmpty()) {
-
-				results = restTemplate.exchange(url, HttpMethod.POST, httpEntity, JsonNode.class);
-				log.info("AdobeOAuth:::results: {}", results);
-				adobeOAuthDto = convertJsonToObj(results.getBody());
-				log.info("AdobeOAuth:::statusCode: {} result: :{}", results.getStatusCode(), results);
-
-			} else {
-				adobeOAuthDto = getOauthAccessTokenFromRefreshToken(oauthRefreshToken);
-				log.info("Get Access Token Used by Refresh Token :::{}", adobeOAuthDto);
-			}
-
-		} catch (RuntimeException runtimeException) {
-			log.info("Exception insdie the:::{}", runtimeException);
-			throw new CBSApplicationException(runtimeException.getLocalizedMessage());
-		}
-
-		saveOrUpdateAdobeKeys(adobeOAuthDto);
-		return adobeOAuthDto;
-
-	}
-
-	private MultiValueMap<String, String> getRequestBody(String refreshToken) {
-
-		Map<String, Config> adobeKeys = getAdobeKeyDetails(ADOBE);
-
-		MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
-
-		body.add(REFRESH_TOKEN, refreshToken);
-		body.add(CLIENT_ID, adobeKeys.get(ADOBE_CLIENT_ID).getKeyValue());
-		body.add(CLIENT_SECRET, adobeKeys.get(ADOBE_CLIENT_SECRET).getKeyValue());
-		body.add(GRANT_TYPE, REFRESH_TOKEN);
-
-		log.info("Refresh Token Body ::: {} ", body);
-		return body;
-	}
-
-	private MultiValueMap<String, String> getRequestBody() {
-
-		Map<String, Config> adobeKeys = getAdobeKeyDetails(ADOBE);
-
-		MultiValueMap<String, String> body = new LinkedMultiValueMap<>();
-
-		body.add(CODE, adobeKeys.get(ADOBE_AUTH_CODE).getKeyValue());
-		body.add(CLIENT_ID, adobeKeys.get(ADOBE_CLIENT_ID).getKeyValue());
-		body.add(CLIENT_SECRET, adobeKeys.get(ADOBE_CLIENT_SECRET).getKeyValue());
-		body.add(REDIRECT_URI, adobeKeys.get(ADOBE_REDIRECT_URL).getKeyValue());
-		body.add(GRANT_TYPE, adobeKeys.get(ADOBE_GRANT_TYPE).getKeyValue());
-
-		log.info("Access Token Body:::{}", body);
-
-		return body;
-	}
-
-	private Map<String, Config> getAdobeKeyDetails(String keyName) {
-
-		log.info("keyName:::{}", keyName);
-
-		Map<String, Config> map = null;
-		List<Config> keysList = null;
-
-		keysList = configRepository.findBykeyNameStartingWith(keyName);
-
-		if (CollectionUtils.isEmpty(keysList))
-			return null;
-
-		map = keysList.stream().collect(Collectors.toMap(Config::getKeyName, config -> config));
-
-		return map;
-	}
-
-	private static boolean isNullOrEmptyMap(Map<?, ?> map) {
-
-		return (map == null || map.isEmpty());
-	}
-
-	private static boolean isMapKeyValueEmptyOrNull(Map<?, ?> map, String key) {
-
-		if (map == null || map.isEmpty())
-			return true;
-
-		if (key == null || key.isEmpty())
-			return true;
-
-		Config config = (Config) map.get(key);
-		log.info("Config object config= {} key= {}" + config);
-
-		return (config == null || config.getKeyValue().isEmpty());
-
-	}
+	@Autowired
+	private AdobeUtility adobeUtility;
 
 	public InputStream downloadAgreement(String agreementId) {
 
@@ -328,14 +88,14 @@ public class AdobeSignServiceImpl implements AdobeSignService {
 
 		try {
 
-			String accessToken = getOauthAccessToken();
+			String accessToken = adobeUtility.getOauthAccessToken();
 
 			HttpHeaders headers = new HttpHeaders();
-			headers.add(HttpHeaders.AUTHORIZATION, getOauthAccessToken());
+			headers.add(HttpHeaders.AUTHORIZATION, accessToken);
 
 			HttpEntity<String> httpEntity = new HttpEntity<>(headers);
 
-			String url = getBaseURIForRestAPI(accessToken) + AGREEMENTS_COMBINEDDOCUMENT;
+			String url = adobeUtility.getBaseURIForRestAPI(accessToken) + AGREEMENTS_COMBINEDDOCUMENT;
 
 			UriComponents uriComponents = UriComponentsBuilder.fromUriString(url).build();
 			uriComponents = uriComponents.expand(Collections.singletonMap("agreementId", agreementId));
@@ -356,7 +116,8 @@ public class AdobeSignServiceImpl implements AdobeSignService {
 	}
 
 	@Transactional
-	private void saveOrUpdateAdobeKeys(AdobeOAuthDto oAuth) {
+	@Override
+	public void saveOrUpdateAdobeKeys(AdobeOAuthDto oAuth) {
 
 		Map<String, Config> result = new HashMap<>();
 
@@ -425,8 +186,8 @@ public class AdobeSignServiceImpl implements AdobeSignService {
 			ResponseEntity<JsonNode> result = null;
 			LinkedMultiValueMap<String, Object> body = null;
 
-			String accessToken = getOauthAccessToken();
-			String transientDocUrl = getBaseURIForRestAPI(accessToken) + TRANSIENT_DOCUMENTS_ENDPOINT;
+			String accessToken = adobeUtility.getOauthAccessToken();
+			String transientDocUrl = adobeUtility.getBaseURIForRestAPI(accessToken) + TRANSIENT_DOCUMENTS_ENDPOINT;
 
 			log.info("transientDocUrl={} fileName={}", transientDocUrl, fileName);
 
@@ -461,8 +222,8 @@ public class AdobeSignServiceImpl implements AdobeSignService {
 
 			ResponseEntity<JsonNode> response = null;
 			HttpHeaders header = null;
-			String accessToken = getOauthAccessToken();
-			String agreementsApiUrl = getBaseURIForRestAPI(accessToken) + AGREEMENTS_ENDPOINT;
+			String accessToken = adobeUtility.getOauthAccessToken();
+			String agreementsApiUrl = adobeUtility.getBaseURIForRestAPI(accessToken) + AGREEMENTS_ENDPOINT;
 			String requestBody = createAgrrementRequestBody(transientDocId, bookingRevision);
 
 			header = new HttpHeaders();
@@ -483,17 +244,20 @@ public class AdobeSignServiceImpl implements AdobeSignService {
 	}
 
 	@Override
-	public boolean saveOrUpdateAuthCode(String authcode) {
-		configRepository.findByKeyName(ADOBE_AUTH_CODE).map(c -> {
-			c.setKeyValue(authcode);
-			return configRepository.save(c);
-		}).orElseGet(() -> {
+	public void saveOrUpdateAuthCode(String authcode) {
+		
+		Optional<Config> config = configRepository.findByKeyName(ADOBE_AUTH_CODE);
+		if(config.isPresent()) {
+			Config con = config.get();
+			con.setKeyValue(authcode);
+			configRepository.save(con);
+		}else{
 			Config con = new Config();
 			con.setKeyName(ADOBE_AUTH_CODE);
 			con.setKeyValue(authcode);
-			return configRepository.save(con);
-		});
-		return false;
+			configRepository.save(con);
+		}
+		
 	}
 
 	@SuppressWarnings("unchecked")
